patch-hook: fix possible cast errors
Livepatch callbacks are packed into specific sections
by gcc by the macros defined in kpatch-macros.h. Callbacks
take the form of:

    struct example {
        type *(fn)(struct klp_object *obj);
        char* objname;
    }

However, create-diff-objects.c can't include the kernel livepatch
header because it is compiled separately from the kernel module.
So the solution before was to define the structs in the header file as:

    struct example {
                type *(fn)(void *obj);
                char* objname;
        }

and then cast fn to (type *(fn)(struct klp_object *obj)) if necessary.
This way, the same header file and definitions can be used in
both places.

But! If compiled with the right GCC flags, such as -Wbad-function-cast,
this casting may throw an error during livepatch building, which is what
I am seeing (not sure why, since it was always fine before...).

Solution: Define the callback function parameter as the correct type, which
allows us to get rid of the troublesome casting.

Since we need to know if callbacks are present in kpatch-patch.h, we can move
the definition of HAVE_CALLBACKS to from livepatch-patch-hook.c to kpatch-patch.h.

IMPORTANT: For correctness, this change requires proper ordering of include
statements! kpatch-patch.h must always be included AFTER linux/livepatch.h
(if using linux/livepatch.h).

Signed-off-by: Brennan Lamoreaux <brennan.lamoreaux@broadcom.com>